requirements.txt
pandas
numpy
requests
kafka-python
pyspark
streamlit
